requirements.txt
GitPython
requests
python-dotenv
# httplib2==0.22.0